\name{get.assoc.studies}
\alias{get.assoc.studies}
\title{Obtain all of the studies associated with a platform}
\usage{
get.assoc.studies(platform_id)
}
\description{
Requires is the user to specify a valid platform e.g. GPL570. Returns the associated studies for a platform.}

\examples{
get.assoc.studies("GPL570")
}
