\name{get.clusters.as.frame}
\alias{get.clusters.as.frame}
\title{Convert cluster objects to frames}
\usage{
get.clusters.as.frame(cluster_list)
}
\Arguments{
\item{cluster_list}{list of integers representing the index of experiment sets. treated labelled 1, control 0 and unused X}
}

\description{
Function used in get.experiment.designs to convert the cluster list generated by functions like title.clustering into a dataframe}

\examples{
gset= obtain.gset("GSE84881")
gsm_titles= obtain.gsm.titles(gset)
title_clusters= title.clustering(gsm_titles)
get.clusters.as.frame(title_clusters)
}
