\name{manual.analysis}
\alias{manual.analysis}
\title{Simplifies the manual assigning of groups}
\usage{
manual.analysisfunction(GEO_set,assigned_groups, group_names= NA, path="~/", clustering, meta_data_and_combined= F, limma_or_rankprod= "limma")
}
\description{
Used to manually assign clusters- needs a GEO_set generated by obtain.gset, assigning of groups in a concatenated list e.g. group_1, group_2, group_3, ect as assigned_groups.
Group names is optional- enables user to specify titles for the ouput files and path specifies where the csvs and metadata are saved. meta_data_and_combined boolean when true saves a metadata table regarding the comparisons whilst the combined table saves a csv table containing the sample sources, titles and characteristics. limma_or_rankprod specifies if user wants to use either limma (10.1093/nar/gkv007) or RankProd (10.1093/bioinformatics/btx292) to caculate fold changes.
}
\examples{
gset= obtain.gset("GSE2328")
combined_table= get.combined.table(gset, T) #F1 the function
combined_table #checking all sample data
groups= rep("X",nrow(combined_table))
groups[1:3]= "group_1"
groups[4:6]= "group_2"
groups[7:9]= "group_3"
manual.analysis(GEO_set= gset,assigned_groups= groups, path="~/", meta_data_and_combined= T)
#Check working directory for the associated csvs and metadata.
}
