\name{obtain.charateristics}
\alias{obtain.charateristics}
\title{Get characteristics GEO dataset}
\usage{
obtain.charateristics(GEOset)
}
\Arguments{
\item{GEOset}{GEO dataseries object- use obtain.gset}
}
\description{
Obtain the characteristics features from a GEO dataset}

\examples{
gset= obtain.gset("GSE84881")
obtain.charateristics(gset)
}

